+++ mtg_proxy_printer/model/document_loader.py
@@ -220,7 +220,7 @@ class DocumentLoader(LoaderSignals):
     MIN_SUPPORTED_SQLITE_VERSION = (3, 31, 0)
 
     def __init__(self, document: "Document", db: sqlite3.Connection = None):  # db parameter used by test code
-        super(DocumentLoader, self).__init__(None)
+        super().__init__(None)
         self.document = document
         self.db = db
         self.finished.connect(functools.partial(self.loading_state_changed.emit, False), Qt.ConnectionType.DirectConnection)
@@ -234,7 +234,8 @@ class DocumentLoader(LoaderSignals):
         logger.info(f"Loading document from {file_path} successful.")
         self.document.save_file_path = file_path
 
-    def cancel(self):
+    @staticmethod
+    def cancel():
         for instance in LoaderRunner.INSTANCES:
             if isinstance(instance, LoaderRunner):
                 instance.cancel()
@@ -705,7 +706,7 @@ def _migrate_5_to_6(db: sqlite3.Connection, settings: PageLayoutSettings):
             ("document_name", settings.document_name),
             ("card_bleed", settings.card_bleed),
             ("draw_page_numbers", settings.draw_page_numbers),
-    ])
+        ])
 
 
 def migrate_image_spacing_settings(db: sqlite3.Connection):