+++ mtg_proxy_printer/ui/cache_cleanup_wizard.py
@@ -25,10 +25,12 @@ from PySide6.QtCore import QAbstractTableModel, Qt, QModelIndex, QObject, QBuffe
 from PySide6.QtGui import QIcon, QPixmap
 from PySide6.QtWidgets import QWidget, QWizard, QWizardPage
 
+import mtg_proxy_printer.settings
 from mtg_proxy_printer.natsort import NaturallySortedSortFilterProxyModel
 from mtg_proxy_printer.model.carddb import CardDatabase, Card, MTGSet
 from mtg_proxy_printer.model.imagedb import ImageDatabase, CacheContent as ImageCacheContent, ImageKey
 from mtg_proxy_printer.ui.common import load_ui_from_file, format_size, WizardBase
+from mtg_proxy_printer.units_and_sizes import OptStr
 from mtg_proxy_printer.logger import get_logger
 logger = get_logger(__name__)
 del get_logger
@@ -52,7 +54,7 @@ Orientation = Qt.Orientation
 
 
 @functools.lru_cache(maxsize=256)
-def get_image_for_tooltip_display(path: pathlib.Path) -> str:
+def get_image_for_tooltip_display(path: pathlib.Path, card_name: OptStr = None) -> str:
     scaling_factor = 3
     source = QPixmap(str(path))
     pixmap = source.scaled(
@@ -62,7 +64,8 @@ def get_image_for_tooltip_display(path: pathlib.Path) -> str:
     buffer.open(QIODevice.OpenModeFlag.WriteOnly)
     pixmap.save(buffer, "PNG", quality=100)
     image = buffer.data().toBase64().toStdString()
-    tooltip_text = f'<img src="data:image/png;base64,{image}">'
+    card_name = f'<p style="text-align:center">{card_name}</p><br>' if card_name else ""
+    tooltip_text = f'{card_name}<img src="data:image/png;base64,{image}">'
     return tooltip_text
 
 
@@ -89,12 +92,13 @@ class KnownCardRow:
     size: int
     scryfall_id: str
     path: pathlib.Path
+    preferred_language_name: OptStr
 
     def data(self, column: int, role: ItemDataRole):
         if column == KnownCardColumns.Name and role in (ItemDataRole.DisplayRole, ItemDataRole.EditRole):
             data = self.name
         elif column == KnownCardColumns.Name and role == ItemDataRole.ToolTipRole:
-            data = get_image_for_tooltip_display(self.path)
+            data = get_image_for_tooltip_display(self.path, self.preferred_language_name)
         elif column == KnownCardColumns.Set:
             data = self.set.data(role)
         elif column == KnownCardColumns.CollectorNumber and role in (ItemDataRole.DisplayRole, ItemDataRole.EditRole):
@@ -142,8 +146,10 @@ class KnownCardImageModel(QAbstractTableModel):
         KnownCardColumns.FilesystemPath: "Path",
     }
 
-    def __init__(self, parent: QObject = None):
-        super(KnownCardImageModel, self).__init__(parent)
+    def __init__(self, card_db: CardDatabase, parent: QObject = None):
+        super().__init__(parent)
+        self.card_db = card_db
+        self.preferred_language: str = mtg_proxy_printer.settings.settings["images"]["preferred-language"]
         self._data: typing.List[KnownCardRow] = []
 
     def rowCount(self, parent: QModelIndex = INVALID_INDEX) -> int:
@@ -157,7 +163,7 @@ class KnownCardImageModel(QAbstractTableModel):
                 and orientation == Orientation.Horizontal \
                 and 0 <= section < self.columnCount():
             return self.header_data[section]
-        return super(KnownCardImageModel, self).headerData(section, orientation, role)
+        return super().headerData(section, orientation, role)
 
     def data(self, index: QModelIndex, role: ItemDataRole = None) -> typing.Any:
         if 0 <= index.row() <= self.rowCount() and 0 <= index.column() < self.columnCount():
@@ -169,9 +175,14 @@ class KnownCardImageModel(QAbstractTableModel):
         position = self.rowCount()
         self.beginInsertRows(INVALID_INDEX, position, position)
         size_bytes = image.absolute_path.stat().st_size
+        if card.language != self.preferred_language:
+            preferred_name = self.card_db.translate_card_name(card, self.preferred_language, True)
+        else:
+            preferred_name = None
         row = KnownCardRow(
             card.name, card.set, card.collector_number, is_hidden,
             image.is_front, image.is_high_resolution, size_bytes, card.scryfall_id, image.absolute_path,
+            preferred_name
         )
         self._data.append(row)
         self.endInsertRows()
@@ -249,7 +260,7 @@ class UnknownCardImageModel(QAbstractTableModel):
     }
 
     def __init__(self, parent: QObject = None):
-        super(UnknownCardImageModel, self).__init__(parent)
+        super().__init__(parent)
         self._data: typing.List[UnknownCardRow] = []
 
     def rowCount(self, parent: QModelIndex = INVALID_INDEX) -> int:
@@ -263,7 +274,7 @@ class UnknownCardImageModel(QAbstractTableModel):
                 and orientation == Orientation.Horizontal \
                 and 0 <= section < self.columnCount():
             return self.header_data[section]
-        return super(UnknownCardImageModel, self).headerData(section, orientation, role)
+        return super().headerData(section, orientation, role)
 
     def data(self, index: QModelIndex, role: ItemDataRole = None) -> typing.Any:
         if 0 <= index.row() < self.rowCount():
@@ -287,7 +298,7 @@ class UnknownCardImageModel(QAbstractTableModel):
 class FilterSetupPage(QWizardPage):
 
     def __init__(self, parent: QWidget = None):
-        super(FilterSetupPage, self).__init__(parent)
+        super().__init__(parent)
         self.ui = Ui_FilterSetupPage()
         self.ui.setupUi(self)
         self.registerField("remove-everything-enabled", self.ui.delete_everything_checkbox)
@@ -302,12 +313,12 @@ class FilterSetupPage(QWizardPage):
 class CardFilterPage(QWizardPage):
 
     def __init__(self, card_db: CardDatabase, image_db: ImageDatabase, parent: QWidget = None):
-        super(CardFilterPage, self).__init__(parent)
+        super().__init__(parent)
         self.ui = Ui_CardFilterPage()
         self.ui.setupUi(self)
         self.card_db = card_db
         self.image_db = image_db
-        self.card_image_model = KnownCardImageModel(parent=self)
+        self.card_image_model = KnownCardImageModel(card_db, self)
         self.card_image_sort_model = self._setup_card_image_sort_model(self.card_image_model)
         self._setup_card_image_view(self.card_image_sort_model)
         self.unknown_image_model = UnknownCardImageModel(parent=self)
@@ -339,7 +350,7 @@ class CardFilterPage(QWizardPage):
             view.setColumnWidth(column, new_size)
 
     def initializePage(self) -> None:
-        super(CardFilterPage, self).initializePage()
+        super().initializePage()
         images = self.image_db.read_disk_cache_content()
         visible, hidden, unknown = self.card_db.get_all_cards_from_image_cache(images)
         for card, key in visible:
@@ -388,7 +399,7 @@ class CardFilterPage(QWizardPage):
             )
 
     def cleanupPage(self) -> None:
-        super(CardFilterPage, self).cleanupPage()
+        super().cleanupPage()
         self.card_image_model.clear()
         self.unknown_image_model.clear()
 
@@ -408,13 +419,13 @@ class CardFilterPage(QWizardPage):
             for index in self.ui.card_image_view.selectedIndexes() if not index.column()
         ]
         self.setField("selected-images", selected_images)
-        return super(CardFilterPage, self).validatePage()
+        return super().validatePage()
 
 
 class SummaryPage(QWizardPage):
 
     def __init__(self, parent: QWidget = None):
-        super(SummaryPage, self).__init__(parent)
+        super().__init__(parent)
         self.ui = Ui_SummaryPage()
         self.ui.setupUi(self)
         logger.info(f"Created {self.__class__.__name__} instance.")
@@ -446,7 +457,7 @@ class CacheCleanupWizard(WizardBase):
         logger.info(f"Created {self.__class__.__name__} instance.")
 
     def accept(self) -> None:
-        super(CacheCleanupWizard, self).accept()
+        super().accept()
         logger.info("User accepted the wizard, deleting entries from the cache.")
         self.image_db.delete_disk_cache_entries((
             ImageKey(scryfall_id, is_front, is_high_resolution)
@@ -455,7 +466,7 @@ class CacheCleanupWizard(WizardBase):
         self._clear_tooltip_cache()
 
     def reject(self) -> None:
-        super(CacheCleanupWizard, self).reject()
+        super().reject()
         logger.info("User canceled the cache cleanup.")
         self._clear_tooltip_cache()